Proper Cars Web Scraping
Rewrote a lot of the logic to properly web scape the website. An ID
range is defined, and chunk size. Each chunk a csv file is made storing
all the successful attempts to read the page. Logic has been added to
make the web requests low intensity, and appear random.